+++ src/ladb_opencutlist/ruby/model/attributes/material_attributes.rb
@@ -14,7 +14,7 @@
     TYPE_EDGE = 4
     TYPE_ACCESSORY = 5
 
-    DEFAULTS = {
+    NATIVES = {
         TYPE_UNKNOWN => {
             :thickness => '0',
             :length_increase => '0',
@@ -31,12 +31,12 @@
         },
         TYPE_SOLID_WOOD => {
             :thickness => '0',
-            :length_increase => '50mm',
-            :width_increase => '5mm',
-            :thickness_increase => '5mm',
+            :length_increase => { :metric => '50mm', :imperial => '50mm' },
+            :width_increase => { :metric => '5mm', :imperial => '5mm' },
+            :thickness_increase => { :metric => '5mm', :imperial => '5mm' },
             :std_lengths => '',
             :std_widths => '',
-            :std_thicknesses => '18mm;27mm;35mm;45mm;64mm;80mm;100mm',
+            :std_thicknesses => { :metric => '18mm;27mm;35mm;45mm;64mm;80mm;100mm', :imperial => '18mm;27mm;35mm;45mm;64mm;80mm;100mm' },
             :std_sections => '',
             :std_sizes => '',
             :grained => true,
@@ -50,7 +50,7 @@
             :thickness_increase => '0',
             :std_lengths => '',
             :std_widths => '',
-            :std_thicknesses => '5mm;8mm;10mm;15mm;18mm;22mm',
+            :std_thicknesses => { :metric => '5mm;8mm;10mm;15mm;18mm;22mm', :imperial => '5mm;8mm;10mm;15mm;18mm;22mm' },
             :std_sections => '',
             :std_sizes => '',
             :grained => false,
@@ -59,25 +59,25 @@
         },
         TYPE_DIMENSIONAL => {
             :thickness => '0',
-            :length_increase => '50mm',
+            :length_increase => { :metric => '50mm', :imperial => '50mm' },
             :width_increase => '0',
             :thickness_increase => '0',
-            :std_lengths => '2400mm',
+            :std_lengths => { :metric => '2400mm', :imperial => '2400mm' },
             :std_widths => '',
             :std_thicknesses => '',
-            :std_sections => '30mm x 40mm;40mm x 50mm',
+            :std_sections => { :metric => '30mm x 40mm;40mm x 50mm', :imperial => '30mm x 40mm;40mm x 50mm' },
             :std_sizes => '',
             :grained => false,
             :edge_decremented => false,
             :volumic_mass => '',
         },
         TYPE_EDGE => {
-            :thickness => '2mm',
-            :length_increase => '50mm',
+            :thickness => { :metric => '2mm', :imperial => '2mm' },
+            :length_increase => { :metric => '50mm', :imperial => '50mm' },
             :width_increase => '0',
             :thickness_increase => '0',
             :std_lengths => '',
-            :std_widths => '23mm;33mm;43mm',
+            :std_widths => { :metric => '23mm;33mm;43mm', :imperial => '23mm;33mm;43mm' },
             :std_thicknesses => '',
             :std_sections => '',
             :std_sizes => '',
@@ -203,6 +203,17 @@
       end
     end
 
+    def self.get_native_value(type, key)
+      unless NATIVES[type].nil? || NATIVES[type][key].nil? || !(NATIVES[type][key].is_a?(Hash))
+        if DimensionUtils.instance.model_unit_is_metric
+          return NATIVES[type][key][:metric] unless NATIVES[type][key][:metric].nil?
+        else
+          return NATIVES[type][key][:imperial] unless NATIVES[type][key][:imperial].nil?
+        end
+      end
+      NATIVES[type][key] unless NATIVES[type][key].is_a?(Hash)
+    end
+
     # -----
 
     def uuid
@@ -223,7 +234,7 @@
       when TYPE_EDGE
         @thickness
       else
-        get_default(:thickness)
+        get_native_value(:thickness)
       end
     end
 
@@ -236,7 +247,7 @@
         when TYPE_SOLID_WOOD, TYPE_SHEET_GOOD, TYPE_DIMENSIONAL, TYPE_EDGE
           @length_increase
         else
-          get_default(:length_increase)
+          get_native_value(:length_increase)
       end
     end
 
@@ -249,7 +260,7 @@
         when TYPE_SOLID_WOOD, TYPE_SHEET_GOOD
           @width_increase
         else
-          get_default(:width_increase)
+          get_native_value(:width_increase)
       end
     end
 
@@ -262,7 +273,7 @@
         when TYPE_SOLID_WOOD
           @thickness_increase
         else
-          get_default(:thickness_increase)
+          get_native_value(:thickness_increase)
       end
     end
 
@@ -275,7 +286,7 @@
       when TYPE_DIMENSIONAL
         @std_lengths
       else
-        get_default(:std_lengths)
+        get_native_value(:std_lengths)
       end
     end
 
@@ -297,7 +308,7 @@
       when TYPE_EDGE
         @std_widths
       else
-        get_default(:std_widths)
+        get_native_value(:std_widths)
       end
     end
 
@@ -319,7 +330,7 @@
         when TYPE_SOLID_WOOD, TYPE_SHEET_GOOD
           @std_thicknesses
         else
-          get_default(:std_thicknesses)
+          get_native_value(:std_thicknesses)
       end
     end
 
@@ -341,7 +352,7 @@
         when TYPE_DIMENSIONAL
           @std_sections
         else
-          get_default(:std_sections)
+          get_native_value(:std_sections)
       end
     end
 
@@ -362,7 +373,7 @@
         when TYPE_SHEET_GOOD
           @std_sizes
         else
-          get_default(:@std_sizes)
+          get_native_value(:@std_sizes)
       end
     end
 
@@ -396,17 +407,17 @@
         end
 
         @type = MaterialAttributes.valid_type(Plugin.instance.get_attribute(@material, 'type', TYPE_UNKNOWN))
-        @thickness = Plugin.instance.get_attribute(@material, 'thickness', get_default(:thickness))
-        @length_increase = Plugin.instance.get_attribute(@material, 'length_increase', get_default(:length_increase))
-        @width_increase = Plugin.instance.get_attribute(@material, 'width_increase', get_default(:width_increase))
-        @thickness_increase = Plugin.instance.get_attribute(@material, 'thickness_increase', get_default(:thickness_increase))
-        @std_lengths = Plugin.instance.get_attribute(@material, 'std_lengths', get_default(:std_lengths))
-        @std_widths = Plugin.instance.get_attribute(@material, 'std_widths', get_default(:std_widths))
-        @std_thicknesses = Plugin.instance.get_attribute(@material, 'std_thicknesses', get_default(:std_thicknesses))
-        @std_sections = Plugin.instance.get_attribute(@material, 'std_sections', get_default(:std_sections))
-        @std_sizes = Plugin.instance.get_attribute(@material, 'std_sizes', get_default(:std_sizes))
-        @grained = Plugin.instance.get_attribute(@material, 'grained', get_default(:grained))
-        @edge_decremented = Plugin.instance.get_attribute(@material, 'edge_decremented', get_default(:edge_decremented))
+        @thickness = Plugin.instance.get_attribute(@material, 'thickness', get_native_value(@type, :thickness))
+        @length_increase = Plugin.instance.get_attribute(@material, 'length_increase', get_native_value(@type, :length_increase))
+        @width_increase = Plugin.instance.get_attribute(@material, 'width_increase', get_native_value(@type, :width_increase))
+        @thickness_increase = Plugin.instance.get_attribute(@material, 'thickness_increase', get_native_value(@type, :thickness_increase))
+        @std_lengths = Plugin.instance.get_attribute(@material, 'std_lengths', get_native_value(@type, :std_lengths))
+        @std_widths = Plugin.instance.get_attribute(@material, 'std_widths', get_native_value(@type, :std_widths))
+        @std_thicknesses = Plugin.instance.get_attribute(@material, 'std_thicknesses', get_native_value(@type, :std_thicknesses))
+        @std_sections = Plugin.instance.get_attribute(@material, 'std_sections', get_native_value(@type, :std_sections))
+        @std_sizes = Plugin.instance.get_attribute(@material, 'std_sizes', get_native_value(@type, :std_sizes))
+        @grained = Plugin.instance.get_attribute(@material, 'grained', get_native_value(@type, :grained))
+        @edge_decremented = Plugin.instance.get_attribute(@material, 'edge_decremented', get_native_value(@type, :edge_decremented))
       else
         @type = TYPE_UNKNOWN
       end
@@ -434,10 +445,6 @@
         @material.set_attribute(Plugin::ATTRIBUTE_DICTIONARY, 'edge_decremented', @edge_decremented)
         @material.set_attribute(Plugin::ATTRIBUTE_DICTIONARY, 'volumic_mass', @volumic_mass)
       end
-    end
-
-    def get_default(key)
-      DEFAULTS[@type][key] unless DEFAULTS[@type].nil?
     end
 
   end